Класс TableFileStream вынесен в отдельные файлы
