+++ src/ctool1cd/tool1cd.cbp
@@ -71,6 +71,8 @@
 		<Unit filename="PrecompiledHeaders.h" />
 		<Unit filename="Table.cpp" />
 		<Unit filename="Table.h" />
+		<Unit filename="TableFileStream.cpp" />
+		<Unit filename="TableFileStream.h" />
 		<Unit filename="TableFiles.cpp" />
 		<Unit filename="TableFiles.h" />
 		<Unit filename="TempStream.cpp" />
